console: finish bulk of changeover to Mithril
Create stateForReplica global for tracking replica state, and make support functions to query it and update it.

Use stateForReplica instead of CSS classes to store information about the replicas (including warning level and whether they should be hidden).

The use of maximumWarningSeverity may need more work and testing relative to when replica_status is set and when CSS is styled. Ideally that global should be removed.