console: add provisional fix for maximumWarningSeverity calculation timing
This fixes an issue where sometimes warning level could be out-of-date due to it being used before the replica drawing routines calculated it using a replica_status global.

This fix is inefficient because it essentially draws replicas an extra time on updates -- and ideally it could be rethought in the process or removing the replica_status global which is modified as a side effect of rendering.